+++ bukkit/changelog.md
@@ -1,1 +1,1 @@
-- Fixed error when port is already in use
+- Fixed server rejecting client connection